Moved settings & registry tweaks to here.
I moved the scripts to set up registry values, windows preferences, and
other stuff from the BoxStarter setup to here.  This way they can be
tweaked and run more often as needed.  Re-running the BoxStarter script
is too cumbersome.  The BoxStarter script instead just focuses on
bootstrapping the machine and getting it ready for me to pull my
WinFiles.