Removing a few apps from auto install
Both pia and rainmeter show dialogs and halt the installation process.
To keep the scripts as automated and hands free as possible, removing
them (to be installed manually instead).  Also, put the Store version of
Spotify back as the non-store one can't be installed from an
administrator shell and blocks the installation.