Updated Windows Terminal settings to support 0.6+
The 0.6 release (a.k.a. 1910 release) of Windows Terminal made a big
change to how the settings are managed.  The settings file now gets
merged with a "default" grouping of settings that the app delivers and
updates.  This means that the user settings can be much shorter and only
need to supply overrides and customizations.