feat: Part 2 of WinGet migration.
This set is mostly just a pass through everything that is not about
application installation to ensure they are in-line with the other
changes.  Tweaked a number of things along the way.  This also leaves
the branch in non-fully usable state.  Before a merge to main I will
have all WinGet updates in place.